Move versioning types into versioning folder
